personal_fianance_tracker/requirements.txt
python-dotenv
rich
google-generativeai
sqlmodel
typer
pandas
plotext
streamlit